Create person again again

+++ CourseManagement/src/main/java/DAL/Teacher/TeacherDAL.java
@@ -14,7 +14,6 @@ public class TeacherDAL extends MyDatabaseManager {
         TeacherDAL.connectDB();
     }
 
-    //2 layer
     public ArrayList readTeacher() throws SQLException {
         String query = "SELECT * FROM Person WHERE HireDate >0";
         ResultSet rs = TeacherDAL.doReadQuery(query);
@@ -99,14 +98,63 @@ public class TeacherDAL extends MyDatabaseManager {
         }
         return list;
     }
+    
+    public int getPersonFromCourseInstructor(int personID) throws SQLException {
 
-    public int deleteTeacher(int personID) throws SQLException {
-        String query = "DELETE FROM Person WHERE PersonID = ?";
+        String query = "SELECT PersonID FROM courseinstructor WHERE PersonID = ? ";
+
         PreparedStatement p = TeacherDAL.getConnection().prepareStatement(query);
         p.setInt(1, personID);
-        int result = p.executeUpdate();
+        ResultSet rs = p.executeQuery();
+        List list = new ArrayList();
 
-        return result;
+        if (rs != null) {
+
+            while (rs.next()) {
+                Teacher s = new Teacher(); //CourseInstructor s = new CourseInstructor(); mới đúng
+                s.setPersonID(rs.getInt("PersonID"));
+                list.add(s);
+            }
+        }
+        if(list.isEmpty()){
+            return 0;
+        }
+        return 1;
+    }
+
+    public int getPersonFromStudentGrade(int personID) throws SQLException {
+
+        String query = "SELECT StudentID FROM studentgrade WHERE StudentID = ?";
+
+        PreparedStatement p = TeacherDAL.getConnection().prepareStatement(query);
+        p.setInt(1, personID);
+        ResultSet rs = p.executeQuery();
+        List list = new ArrayList();
+
+        if (rs != null) {
+
+            while (rs.next()) {
+                Teacher s = new Teacher(); //StudentGrade s = new StudentGrade(); mới đúng
+                s.setPersonID(rs.getInt("StudentID"));
+                list.add(s);
+            }
+        }
+        if(list.isEmpty()){
+            return 0;
+        }
+        return 1;
+    }
+    
+    public int deleteTeacher(int personID) throws SQLException {
+        if (getPersonFromCourseInstructor(personID) == 1 || getPersonFromStudentGrade(personID) == 1) {
+            return 0;
+        } else {
+            String query = "DELETE FROM Person WHERE PersonID = ?";
+            PreparedStatement p = TeacherDAL.getConnection().prepareStatement(query);
+            p.setInt(1, personID);
+            p.executeUpdate();
+            return 1;
+        }
     }
 
     public static void main(String[] args) {
